examples/wyggles/pyproject.toml
[tool.poetry]
name = "wyggles"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
arcade = "^2.6.16"
pycairo = "^1.21.0"
botsley = {path = "../..", develop = true}
numpy = "^1.23.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
